+++ src/main/java/com/rollingpaper/ggeujeogggeujeog/user/presentation/UserController.java
@@ -9,15 +9,13 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.rollingpaper.ggeujeogggeujeog.user.application.UserService;
-import com.rollingpaper.ggeujeogggeujeog.user.presentation.dto.SignUpRequestDto;
-import com.rollingpaper.ggeujeogggeujeog.user.presentation.dto.UserProfileResponseDto;
 import com.rollingpaper.ggeujeogggeujeog.user.presentation.dto.UserUpdateRequestDto;
+import com.rollingpaper.ggeujeogggeujeog.user.application.UserService;
+import com.rollingpaper.ggeujeogggeujeog.user.presentation.dto.UserProfileResponseDto;
 
 import lombok.RequiredArgsConstructor;
 
@@ -27,14 +25,6 @@ import lombok.RequiredArgsConstructor;
 public class UserController {
 
 	private final UserService userService;
-
-	@PostMapping
-	public ResponseEntity<Void> signUp(
-		@RequestBody @Valid SignUpRequestDto dto
-	) {
-		userService.register(dto);
-		return ResponseEntity.status(OK).build();
-	}
 
 	@PatchMapping("/{id}")
 	public ResponseEntity<Void> update(
