+++ src/test/java/com/rollingpaper/ggeujeogggeujeog/authentication/application/SessionLoginServiceTest.java
@@ -16,7 +16,9 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import com.rollingpaper.ggeujeogggeujeog.authentication.presentation.dto.SignInRequestDto;
+import com.rollingpaper.ggeujeogggeujeog.authentication.presentation.dto.SignUpRequestDto;
 import com.rollingpaper.ggeujeogggeujeog.common.util.PasswordEncoder;
+import com.rollingpaper.ggeujeogggeujeog.user.exception.DuplicatedEmailException;
 import com.rollingpaper.ggeujeogggeujeog.user.exception.NoSuchUserException;
 import com.rollingpaper.ggeujeogggeujeog.user.infrastructure.UserMapper;
 
@@ -71,5 +73,38 @@ class SessionLoginServiceTest {
 
 		//then
 		assertThrows(IllegalArgumentException.class, () -> sessionLoginService.signIn(dto));
+	}
+
+	@Test
+	@DisplayName("중복된 이메일이 없다면 회원가입에 성공한다")
+	void register() {
+		//given
+		SignUpRequestDto dto = SignUpRequestDto.builder()
+			.email(TestUser.USER1.getEmail())
+			.password(TestUser.USER1.getPassword())
+			.nickname(TestUser.USER1.getNickname())
+			.build();
+		given(userMapper.findByEmail(any())).willReturn(Optional.ofNullable(null));
+
+		//when
+		sessionLoginService.register(dto);
+
+		//then
+		then(userMapper).should(times(1)).save(any());
+	}
+
+	@Test
+	@DisplayName("중복된 이메일이 있다면 회원가입에 실패한다")
+	void registerWithDuplicatedEmail() {
+		//given
+		SignUpRequestDto dto = SignUpRequestDto.builder()
+			.email(TestUser.USER1.getEmail())
+			.password(TestUser.USER1.getPassword())
+			.nickname(TestUser.USER1.getNickname())
+			.build();
+		given(userMapper.findByEmail(any())).willReturn(Optional.ofNullable(SignUpRequestDto.toEntity(dto)));
+
+		//then
+		assertThrows(DuplicatedEmailException.class, () -> sessionLoginService.register(dto));
 	}
 }
